Extract KeyframeRecord and use a dynamic import
Resolves #31
