+++ index.mjs
@@ -9,4 +9,5 @@ export * from "./class/level/drone/Drone.mjs"
 export * from "./class/level/drone/DroneTransmitter.mjs"
 export * from "./class/level/drone/Ego.mjs"
 export * from "./class/level/changeRecord/ChangeRecord.mjs"
+export * from "./class/level/changeRecord/KeyframeRecord.mjs"
 export * from "./class/level/changeRecord/NullChangeRecord.mjs"
